can delete admin

+++ controlers/del_user.php
@@ -6,12 +6,16 @@ $users = new Users;
 if (isset($_GET["id"]) and !empty($_GET["id"])) {
     $id = $_GET["id"];
 
-    try {
-        //code...
-        $users->delete($id);
-        $res = "supprime";
-    } catch (EXCEPTION $e) {
-        $res = "ERREUR" . $e->getMessage();
+    if ($id != 1 and $id != 0) {
+        try {
+            //code...
+            $users->delete($id);
+            $res = "supprime";
+        } catch (EXCEPTION $e) {
+            $res = "ERREUR" . $e->getMessage();
+        }
+    } else {
+        $res = "impossible de supprimer l'admin";
     }
     echo json_encode($res);
 }
